Do not derive PartialEq for Error
Our Error type is going to contain `io::Error` once we start doing I/O,
which is not `PartialEq`, so we won't be able to auto derive it.
Instead, we use `assert_matches!` (from the `assert_matches` crate) to
check the overall structure of our errors (i.e., what variant they are)
and then use `failure::Fail`'s ability to downcast to the concrete error
type and check that instead.